Relocate supervisord to the flavored images and make base much more explicit
The base image is not really sensible for end-users in any way and only serves
as a means of consolidating system dependencies before the evergreen-client and
remaining distribution is installed.

This change makes this relationship much more explicit than it was previously,
since the backend would reject flavor-less registrations anyways.

Once this change is merged, jenkins/evergreen:latest should be deleted from
Dockerhub

+++ distribution/tests/offline-tests.sh
@@ -70,17 +70,20 @@ test_jep_307() {
 
 # JENKINS-53059
 test_INSECURE_SHOW_ADMIN_PASSWORD_can_be_unset() {
-  result=$( docker run --rm jenkins/evergreen:$ENVIRONMENT /evergreen/scripts/jenkins-evergreen.sh )
-  # Expected to fail, but because war is missing. Admin password must have been generated,
-  # and final expected error line is "Error: Unable to access jarfile /evergreen/jenkins/home/jenkins.war"
-  assertNotEquals "Should have failed to start up (war is absent)" "0" "$?"
+  # Avoid executing this for the base image
+  if [ ! -z "${ENVIRONMENT}" ]; then
+    result=$( docker run --rm jenkins/evergreen:$ENVIRONMENT /evergreen/scripts/jenkins-evergreen.sh )
+    # Expected to fail, but because war is missing. Admin password must have been generated,
+    # and final expected error line is "Error: Unable to access jarfile /evergreen/jenkins/home/jenkins.war"
+    assertNotEquals "Should have failed to start up (war is absent)" "0" "$?"
 
-  echo "$result" | grep -v '[admin password] '
-  assertEquals "Line with generated password should not have been found" "0" "$?"
+    echo "$result" | grep -v '[admin password] '
+    assertEquals "Line with generated password should not have been found" "0" "$?"
 
-  result=$( docker logs "$container_under_test" |
-    grep -e '^Error: Unable to access jarfile /evergreen/jenkins/home/jenkins.war')
-  assertEquals "Should not have been able to start jenkins.war" "0" "$?"
+    result=$( docker logs "$container_under_test" |
+        grep -e '^Error: Unable to access jarfile /evergreen/jenkins/home/jenkins.war')
+    assertEquals "Should not have been able to start jenkins.war" "0" "$?"
+  fi;
 }
 
 test_docker_flavor_custom_supervisordconf() {
